requirements.txt
# Audio Processing
pydub==0.25.1
soundfile==0.12.1
ffmpeg-python==0.2.0
essentia==2.1b5  # Add Essentia package (make sure to pick the most stable version)
numpy==1.22.4  # Specify a compatible version for numpy to avoid conflicts with numba

# Web Development
Flask==2.3.3
gunicorn==23.0.0

# Utility Libraries
requests==2.28.1
python-dotenv==1.0.1
tqdm==4.65.1
Werkzeug>=2.3.7  # Ensure compatibility with the latest Flask
